+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v2.4.6
+
+* Added: Optional benchmarks unit test
+
 # v2.4.5
 
 * Added: Vertex.query limited by connected edge labels